Refactor profile configuration
This commit refactors the profile configuration to allow for easier use
when you have full domains, as opposed to the last configuration which
required separate subdomains for each root domain.

The Nix module has been updated accordingly, and the tests have been
updated to reflect the new configuration.

This is a breaking change, as the configuration format has changed.